new algo

+++ 394199/lock.cpp
@@ -5,11 +5,11 @@ bool lock_acquire(struct lock_t* lock) {
     int bounded_try = 1;
     do {
         // We use the lock_word's LSB for the lock state and the rest for the version number
-        int current = lock->version.load();
+        uint64_t current = lock->version.load();
         
         // Prepare desired value: lock bit is 1 (locked) and version stays the same
-        int expected = current & ~1; // we want to make sure the expected is not locked
-        int desired = current | 1;  // Set the LSB to 1 (locked)
+        uint64_t expected = current & ~1; // we want to make sure the expected is not locked
+        uint64_t desired = current | 1;  // Set the LSB to 1 (locked)
 
         // Perform CAS (Compare-And-Swap)
         while (!lock->version.compare_exchange_strong(expected, desired)) {
@@ -31,7 +31,7 @@ void lock_release(struct lock_t* lock) {
     lock->version.store(lock->version.load() & ~1);
 }
 
-int get_version(struct lock_t* lock){
+uint64_t get_version(struct lock_t* lock){
     return lock->version.load() >> 1;
 }
